Refactor FXIOS-6594 [v116] Reorganize General & Account settings (#14791)
* Reorganize General & Account settings

* Apply suggestions from code review


* Indentation changes

* swiftlint

* Fix issue with duplicate setting

---------
