+++ Client/Frontend/Settings/Main/Account/ChinaSyncServiceSetting.swift
@@ -1,0 +1,69 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Account
+import Common
+import Foundation
+import Shared
+
+class ChinaSyncServiceSetting: Setting {
+    override var accessoryType: UITableViewCell.AccessoryType { return .none }
+    var prefs: Prefs { return profile.prefs }
+    let prefKey = PrefsKeys.KeyEnableChinaSyncService
+    let profile: Profile
+    let settings: UIViewController
+
+    override var hidden: Bool { return !AppInfo.isChinaEdition }
+
+    override var title: NSAttributedString? {
+        return NSAttributedString(string: "本地同步服务",
+                                  attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
+    }
+
+    override var status: NSAttributedString? {
+        return NSAttributedString(string: "禁用后使用全球服务同步数据",
+                                  attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textSecondary])
+    }
+
+    init(settings: SettingsTableViewController) {
+        self.profile = settings.profile
+        self.settings = settings
+    }
+
+    override func onConfigureCell(_ cell: UITableViewCell, theme: Theme) {
+        super.onConfigureCell(cell, theme: theme)
+        let control = UISwitch()
+        control.onTintColor = theme.colors.actionPrimary
+        control.addTarget(self, action: #selector(switchValueChanged), for: .valueChanged)
+        control.isOn = prefs.boolForKey(prefKey) ?? AppInfo.isChinaEdition
+        cell.accessoryView = control
+        cell.selectionStyle = .none
+    }
+
+    @objc
+    func switchValueChanged(_ toggle: UISwitch) {
+        TelemetryWrapper.recordEvent(category: .action, method: .tap, object: .chinaServerSwitch)
+        guard profile.rustFxA.hasAccount() else {
+            prefs.setObject(toggle.isOn, forKey: prefKey)
+            RustFirefoxAccounts.reconfig(prefs: profile.prefs)
+            return
+        }
+
+        // Show confirmation dialog for the user to sign out of FxA
+
+        let msg = "更改此设置后，再次登录您的帐户" // "Sign-in again to your account after changing this setting"
+        let alert = UIAlertController(title: "", message: msg, preferredStyle: .alert)
+        let okString = UIAlertAction(title: .OKString, style: .default) { _ in
+            self.prefs.setObject(toggle.isOn, forKey: self.prefKey)
+            self.profile.removeAccount()
+            RustFirefoxAccounts.reconfig(prefs: self.profile.prefs)
+        }
+        let cancel = UIAlertAction(title: .CancelString, style: .default) { _ in
+            toggle.setOn(!toggle.isOn, animated: true)
+        }
+        alert.addAction(okString)
+        alert.addAction(cancel)
+        settings.present(alert, animated: true)
+    }
+}
