+++ Client/Frontend/Settings/Main/SettingDisclosureUtility.swift
@@ -1,0 +1,16 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+import Shared
+
+struct SettingDisclosureUtility {
+    static func buildDisclosureIndicator(theme: Theme) -> UIImageView {
+        let disclosureIndicator = UIImageView()
+        disclosureIndicator.image = UIImage(named: ImageIdentifiers.menuChevron)?.withRenderingMode(.alwaysTemplate).imageFlippedForRightToLeftLayoutDirection()
+        disclosureIndicator.tintColor = theme.colors.actionSecondary
+        disclosureIndicator.sizeToFit()
+        return disclosureIndicator
+    }
+}
